divide codebase into modules, finished app
