+++ modules/book.js
@@ -1,0 +1,9 @@
+class Book {
+  constructor(title, author) {
+    this.title = title;
+    this.author = author;
+    this.id = new Date().valueOf().toString();
+  }
+}
+
+export default Book;
