+++ package.json
@@ -15,5 +15,11 @@
   "bugs": {
     "url": "https://github.com/franklinben23/awesome-books-es6/issues"
   },
-  "homepage": "https://github.com/franklinben23/awesome-books-es6#readme"
+  "homepage": "https://github.com/franklinben23/awesome-books-es6#readme",
+  "devDependencies": {
+    "babel-eslint": "^10.1.0",
+    "eslint": "^7.32.0",
+    "eslint-config-airbnb-base": "^14.2.1",
+    "eslint-plugin-import": "^2.25.4"
+  }
 }
